pyproject.toml
[tool.poetry]
name = "goslinks"
version = "0.0.1"
description = "URL shortener for teams."
authors = ["Lucas Connors <lucas@revolutiontech.ca>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1"
pynamodb = "^4.3"
authlib = "^0.14"
google-auth = "^1.16"
google-api-python-client = "^1.8"
python-dotenv = "^0.13.0"
Flask-Testing = "^0.8.0"
flask-wtf = "^0.14.3"
zappa = "^0.51.0"
loginpass = "^0.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
coverage = "^5.1"
pyupgrade = "^2.4"
isort = "^4.3.21"

[tool.isort]
# https://github.com/psf/black/blob/master/docs/compatible_configs.md#isort
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.black]
line-length = 88
target_version = ["py38"]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
